Big refactoring:
 - lazy values -related stuff is now in a dedicated file
 - lazy tuple items now have a correct id generated even in old pytest 2
 - all mark-related utils are in a dedicated file. Fixed a with issues with marks related with pytest 2 (make_marked_parameter_value, etc.)
 - new parametrization style in `@parametrize`: cartesian product of argnames, argvalues is now correct even when custom marks and lazy values are present
