+++ pytest_cases/fixture_core2.py
@@ -21,6 +21,7 @@ try:  # type hints, python 3+
 except ImportError:
     pass
 
+from .common_pytest_lazy_values import get_lazy_args
 from .common_pytest import get_pytest_parametrize_marks, make_marked_parameter_value, get_param_argnames_as_list, \
     analyze_parameter_set, combine_ids, is_marked_parameter_value, get_marked_parameter_values, pytest_fixture
 from .fixture__creation import get_caller_module, check_name_available, WARN, CHANGE
@@ -408,7 +409,7 @@ def _decorate_fixture_plus(fixture_func,
         # reapply the marks
         for i, marks in enumerate(final_marks):
             if marks is not None:
-                final_values[i] = make_marked_parameter_value(final_values[i], marks=marks)
+                final_values[i] = make_marked_parameter_value((final_values[i],), marks=marks)
     else:
         final_values = list(product(*params_values))
         final_ids = combine_ids(product(*params_ids))
@@ -418,7 +419,7 @@ def _decorate_fixture_plus(fixture_func,
         for i, marks in enumerate(final_marks):
             ms = [m for mm in marks if mm is not None for m in mm]
             if len(ms) > 0:
-                final_values[i] = make_marked_parameter_value(final_values[i], marks=ms)
+                final_values[i] = make_marked_parameter_value((final_values[i],), marks=ms)
 
     if len(final_values) != len(final_ids):
         raise ValueError("Internal error related to fixture parametrization- please report")
@@ -440,8 +441,6 @@ def _decorate_fixture_plus(fixture_func,
 
     # --common routine used below. Fills kwargs with the appropriate names and values from fixture_params
     def _map_arguments(*_args, **_kwargs):
-        # todo better...
-        from .fixture_parametrize_plus import handle_lazy_args
         request = _kwargs['request'] if func_needs_request else _kwargs.pop('request')
 
         # populate the parameters
@@ -452,12 +451,12 @@ def _decorate_fixture_plus(fixture_func,
         for p_names, fixture_param_value in zip(params_names_or_name_combinations, _params):
             if len(p_names) == 1:
                 # a single parameter for that generated fixture (@pytest.mark.parametrize with a single name)
-                _kwargs[p_names[0]] = handle_lazy_args(fixture_param_value)
+                _kwargs[p_names[0]] = get_lazy_args(fixture_param_value)
             else:
                 # several parameters for that generated fixture (@pytest.mark.parametrize with several names)
                 # unpack all of them and inject them in the kwargs
                 for old_p_name, old_p_value in zip(p_names, fixture_param_value):
-                    _kwargs[old_p_name] = handle_lazy_args(old_p_value)
+                    _kwargs[old_p_name] = get_lazy_args(old_p_value)
 
         return _args, _kwargs
