+++ pytest_cases/fixture_parametrize_plus.py
@@ -1,9 +1,7 @@
-from collections import Iterable, namedtuple
-from distutils.version import LooseVersion
-from functools import partial
+from collections import Iterable
 from inspect import isgeneratorfunction
-from itertools import product
 from warnings import warn
+
 
 try:  # python 3.3+
     from inspect import signature, Parameter
@@ -12,6 +10,7 @@ except ImportError:
 
 try:
     from typing import Union, Callable, List, Any, Sequence, Optional  # noqa
+
 except ImportError:
     pass
 
@@ -20,9 +19,10 @@ from makefun import with_signature, remove_signature_parameters, add_signature_p
 
 from .common_mini_six import string_types
 from .common_others import AUTO
-from .common_pytest import get_fixture_name, remove_duplicates, is_marked_parameter_value, mini_idvalset, \
-    get_param_argnames_as_list, extract_parameterset_info, ParameterSet, has_pytest_param, get_pytest_marks_on_function, \
-    transform_marks_into_decorators
+from .common_pytest_marks import has_pytest_param, get_param_argnames_as_list
+from .common_pytest_lazy_values import is_lazy_value, is_lazy, get_lazy_args
+from .common_pytest import get_fixture_name, remove_duplicates, mini_idvalset, is_marked_parameter_value, \
+    extract_parameterset_info, ParameterSet, cart_product_pytest
 
 from .fixture__creation import check_name_available, CHANGE, WARN, get_caller_module
 from .fixture_core1_unions import InvalidParamsList, NOT_USED, UnionFixtureAlternative, _make_fixture_union, \
@@ -60,7 +60,7 @@ def _fixture_product(caller_module,
     if not isinstance(fixtures_or_values, (tuple, set, list)):
         raise TypeError("fixture_product: the `fixtures_or_values` argument should be a tuple, set or list")
     else:
-        has_lazy_vals = any(isinstance(v, lazy_value) for v in fixtures_or_values)
+        has_lazy_vals = any(is_lazy_value(v) for v in fixtures_or_values)
 
     _tuple_size = len(fixtures_or_values)
 
@@ -85,7 +85,7 @@ def _fixture_product(caller_module,
             if fix_at_pos_i is None:
                 # fixed value
                 # note: wouldnt it be almost as efficient but more readable to *always* call handle_lazy_args?
-                yield handle_lazy_args(fixtures_or_values[i]) if has_lazy_vals else fixtures_or_values[i]
+                yield get_lazy_args(fixtures_or_values[i]) if has_lazy_vals else fixtures_or_values[i]
             else:
                 # fixture value
                 yield all_fixtures[fix_at_pos_i]
@@ -126,166 +126,6 @@ class fixture_ref(object):  # noqa
 
     def __init__(self, fixture):
         self.fixture = fixture
-
-
-pytest53 = LooseVersion(pytest.__version__) >= LooseVersion("5.3.0")
-if pytest53:
-    # in the latest versions of pytest, the default _idmaker returns the value of __name__ if it is available,
-    # even if an object is not a class nor a function. So we do not need to use any special trick.
-    _LazyValueBase = object
-else:
-    fake_base = int
-
-    class _LazyValueBase(int, object):
-        """
-        in this older version of pytest, the default _idmaker does *not* return the value of __name__ for
-        objects that are not functions not classes. However it *does* return str(obj) for objects that are
-        instances of bool, int or float. So that's why lazy_value inherits from int.
-        """
-        __slots__ = ()
-
-        def __new__(cls,
-                    valuegetter,  # type: Callable[[], Any]
-                    id=None,      # type: str
-                    marks=()      # type: Sequence
-                    ):
-            """ Inheriting from int is a bit hard in python: we have to override __new__ """
-            obj = fake_base.__new__(cls, 111111)
-            cls.__init__(obj, valuegetter=valuegetter, id=id, marks=marks)
-            return obj
-
-        def __getattribute__(self, item):
-            """Map all default attribute and method access to the ones in object, not in int"""
-            return object.__getattribute__(self, item)
-
-        def __repr__(self):
-            """Magic methods are not intercepted by __getattribute__ and need to be overridden manually.
-            We do not need all of them by at least override this one for easier debugging"""
-            return object.__repr__(self)
-
-
-def _unwrap(obj):
-    """A light copy of _pytest.compat.get_real_func. In our case
-    we do not wish to unwrap the partial nor handle pytest fixture
-    Note: maybe from inspect import unwrap could do the same?
-    """
-    start_obj = obj
-    for i in range(100):
-        # __pytest_wrapped__ is set by @pytest.fixture when wrapping the fixture function
-        # to trigger a warning if it gets called directly instead of by pytest: we don't
-        # want to unwrap further than this otherwise we lose useful wrappings like @mock.patch (#3774)
-        # new_obj = getattr(obj, "__pytest_wrapped__", None)
-        # if isinstance(new_obj, _PytestWrapper):
-        #     obj = new_obj.obj
-        #     break
-        new_obj = getattr(obj, "__wrapped__", None)
-        if new_obj is None:
-            break
-        obj = new_obj
-    else:
-        raise ValueError("could not find real function of {start}\nstopped at {current}".format(
-                start=repr(start_obj), current=repr(obj)
-            )
-        )
-    return obj
-
-
-def partial_to_str(partialfun):
-    """Return a string representation of a partial function, to use in lazy_value ids"""
-    strwds = ", ".join("%s=%s" % (k, v) for k, v in partialfun.keywords.items())
-    if len(partialfun.args) > 0:
-        strargs = ', '.join(str(i) for i in partialfun.args)
-        if len(partialfun.keywords) > 0:
-            strargs = "%s, %s" % (strargs, strwds)
-    else:
-        strargs = strwds
-    return "%s(%s)" % (partialfun.func.__name__, strargs)
-
-
-# noinspection PyPep8Naming
-class lazy_value(_LazyValueBase):
-    """
-    A reference to a value getter, to be used in `parametrize_plus`.
-
-    A `lazy_value` is the same thing than a function-scoped fixture, except that the value getter function is not a
-    fixture and therefore can neither be parametrized nor depend on fixtures. It should have no mandatory argument.
-    """
-    if pytest53:
-        __slots__ = 'valuegetter', '_id', '_marks'
-    else:
-        # we can not define __slots__ since we extend int,
-        # see https://docs.python.org/3/reference/datamodel.html?highlight=__slots__#notes-on-using-slots
-        pass
-
-    # noinspection PyMissingConstructor
-    def __init__(self,
-                 valuegetter,  # type: Callable[[], Any]
-                 id=None,      # type: str
-                 marks=()      # type: Sequence
-                 ):
-        """
-        Creates a reference to a value getter, to be used in `parametrize_plus`.
-
-        A `lazy_value` is the same thing than a function-scoped fixture, except that the value getter function is not a
-        fixture and therefore can neither be parametrized nor depend on fixtures. It should have no mandatory argument.
-
-        Note that a `lazy_value` can be included in a `pytest.param` without problem. In that case the id defined by
-        `pytest.param` will take precedence over the one defined in `lazy_value` if any. The marks, however,
-        will all be kept wherever they are defined.
-
-        :param valuegetter: a callable without mandatory arguments
-        :param id: an optional id. Otherwise `valuegetter.__name__` will be used by default
-        :param marks: optional marks. `valuegetter` marks will also be preserved.
-        """
-        self.valuegetter = valuegetter
-        self._id = id
-        if isinstance(marks, (tuple, list, set)):
-            self._marks = marks
-        else:
-            self._marks = (marks, )
-
-    def get_marks(self, as_decorators=False):
-        """
-        Overrides default implementation to return the marks that are on the case function
-
-        :param as_decorators: when True, the marks will be transformed into MarkDecorators before being
-            returned
-        :return:
-        """
-        valuegetter_marks = get_pytest_marks_on_function(self.valuegetter, as_decorators=as_decorators)
-
-        if self._marks:
-            return transform_marks_into_decorators(self._marks) + valuegetter_marks
-        else:
-            return valuegetter_marks
-
-    def get_id(self):
-        """The id to use in pytest"""
-        if self._id is not None:
-            return self._id
-        else:
-            # default is the __name__ of the value getter
-            _id = getattr(self.valuegetter, '__name__', None)
-            if _id is not None:
-                return _id
-
-            # unwrap and handle partial functions
-            vg = _unwrap(self.valuegetter)
-
-            if isinstance(vg, partial):
-                return partial_to_str(vg)
-            else:
-                return vg.__name__
-
-    if not pytest53:
-        def __str__(self):
-            """in pytest<5.3 we inherit from int so that str(v) is called by pytest _idmaker to get the id"""
-            return self.get_id()
-
-    @property
-    def __name__(self):
-        """for pytest >= 5.3 we override this so that pytest uses it for id"""
-        return self.get_id()
 
 
 # Fix for https://github.com/smarie/python-pytest-cases/issues/71
@@ -502,91 +342,6 @@ def parametrize_plus(argnames=None,
                              hook=hook, debug=debug, **args)
 
 
-def handle_lazy_args(argval):
-    """ Possibly calls the lazy values contained in argval if needed, before returning it"""
-
-    # First handle the general case
-    try:
-        if not isinstance(argval, (lazy_value, LazyTuple, LazyTuple.LazyItem)):
-            return argval
-    except:  # noqa
-        return argval
-
-    # Now the lazy ones
-    if isinstance(argval, lazy_value):
-        return argval.valuegetter()
-    elif isinstance(argval, LazyTuple):
-        return argval.get()
-    elif isinstance(argval, LazyTuple.LazyItem):
-        return argval.get()
-    else:
-        return argval
-
-
-class LazyTuple(object):
-    """
-    A wrapper representing a lazy_value used as a tuple = for several argvalues at once.
-
-     -
-       while not calling the lazy value
-     -
-    """
-    __slots__ = ('value', 'theoretical_size', 'retrieved')
-
-    def __init__(self,
-                 valueref,        # type: Union[lazy_value, Sequence]
-                 theoretical_size  # type: int
-                 ):
-        self.value = valueref
-        self.theoretical_size = theoretical_size
-        self.retrieved = False
-
-    def __len__(self):
-        return self.theoretical_size
-
-    def get_id(self):
-        """return the id to use by pytest"""
-        return self.value.get_id()
-
-    class LazyItem(namedtuple('LazyItem', ('host', 'item'))):
-        def get(self):
-            return self.host.force_getitem(self.item)
-
-    def __getitem__(self, item):
-        """
-        Getting an item in the tuple with self[i] does *not* retrieve the value automatically, but returns
-        a facade (a LazyItem), so that pytest can store this item independently wherever needed, without
-        yet calling the value getter.
-        """
-        if self.retrieved:
-            # this is never called by pytest, but keep it for debugging
-            return self.value[item]
-        else:
-            # do not retrieve yet: return a facade
-            return LazyTuple.LazyItem(self, item)
-
-    def force_getitem(self, item):
-        """ Call the underlying value getter, then return self[i]. """
-        getter = self.value
-        argvalue = self.get()
-        try:
-            return argvalue[item]
-        except TypeError as e:
-            raise ValueError("(lazy_value) The parameter value returned by `%r` is not compliant with the number"
-                             " of argnames in parametrization (%s). A %s-tuple-like was expected. "
-                             "Returned lazy argvalue is %r and argvalue[%s] raised %s: %s"
-                             % (getter.valuegetter, self.theoretical_size, self.theoretical_size,
-                                argvalue, item, e.__class__, e))
-
-    def get(self):
-        """ Call the underlying value getter, then return the tuple (not self) """
-        if not self.retrieved:
-            # retrieve
-            self.value = self.value.valuegetter()
-            self.retrieved = True
-        return self.value
-
-
 class InvalidIdTemplateException(Exception):
     """
     Raised when a string template provided in an `idgen` raises an error
@@ -646,6 +401,9 @@ def _parametrize_plus(argnames=None,
     if len(fixture_indices) == 0:
         if debug:
             print("No fixture reference found. Calling @pytest.mark.parametrize...")
+            print(" - argnames: %s" % initial_argnames)
+            print(" - argvalues: %s" % marked_argvalues)
+            print(" - ids: %s" % ids)
 
         # no fixture reference: shortcut, do as usual (note that the hook wont be called since no fixture is created)
         _decorator = pytest.mark.parametrize(initial_argnames, marked_argvalues, indirect=indirect,
@@ -696,7 +454,7 @@ def _parametrize_plus(argnames=None,
                     raise ValueError("Internal error, please report")
                 if len(argnames) <= 1:
                     argvals = (argvals,)
-                elif isinstance(argvals, LazyTuple):
+                elif is_lazy(argvals):
                     return argvals.get_id()
                 return mini_idvalset(argnames, argvals, idx=_tmp_make_id.i)
 
@@ -1007,18 +765,10 @@ def _get_argnames_argvalues(argnames=None, argvalues=None, **args):
     :return: argnames, argvalues - both guaranteed to be lists
     """
     # handle **args - a dict of {argnames: argvalues}
-    if len(args) < 2:
-        kw_argnames = get_param_argnames_as_list(next(iter(args.keys()))) if len(args) > 0 else []
-        kw_argvalues = list(*args.values())
+    if len(args) > 0:
+        kw_argnames, kw_argvalues = cart_product_pytest(tuple(args.keys()), tuple(args.values()))
     else:
-        kw_argvalues = list(product(*args.values()))
-        kw_argnames = []
-        for i, argname in enumerate(args.keys()):
-            _names = get_param_argnames_as_list(argname)
-            kw_argnames += _names
-            if len(_names) > 1:
-                for j, _argvals in enumerate(kw_argvalues):
-                    kw_argvalues[j] = kw_argvalues[j][:i] + kw_argvalues[j][i] + kw_argvalues[j][i+1:]
+        kw_argnames, kw_argvalues = (), ()
 
     if argnames is None:
         # (1) all {argnames: argvalues} pairs are provided in **args
@@ -1027,6 +777,9 @@ def _get_argnames_argvalues(argnames=None, argvalues=None, **args):
 
         argnames = kw_argnames
         argvalues = kw_argvalues
+        # simplify if needed to comply with pytest.mark.parametrize
+        if len(argnames) == 1:
+            argvalues = [l[0] if not is_marked_parameter_value(l) else l for l in argvalues]
 
     elif isinstance(argnames, string_types):
         # (2) argnames + argvalues, as usual. However **args can also be passed and should be added
@@ -1044,7 +797,10 @@ def _get_argnames_argvalues(argnames=None, argvalues=None, **args):
         # append **args
         if len(kw_argnames) > 0:
             argnames.extend(kw_argnames)
-            argvalues = list(product(argvalues, *kw_argvalues))
+            # TODO check this case
+            # cart_product(*args.values())  # do NOT use `itertools.product` as it fails to handle MarkDecorators
+            # argvalues = list(product(argvalues, *kw_argvalues))
+            argvalues = cart_product_pytest(argvalues, *kw_argvalues)
 
     return argnames, argvalues
 
@@ -1098,7 +854,7 @@ def _process_argvalues(argnames, marked_argvalues, nb_params):
     fixture_indices = []
     if nb_params == 1:
         for i, v in enumerate(argvalues):
-            if isinstance(v, lazy_value):
+            if is_lazy_value(v):
                 # Note: no need to modify the id, it will be ok thanks to the lazy_value class design
                 # handle marks
                 _mks = v.get_marks(as_decorators=True)
@@ -1116,9 +872,9 @@ def _process_argvalues(argnames, marked_argvalues, nb_params):
                 fixture_indices.append((i, None))
     elif nb_params > 1:
         for i, v in enumerate(argvalues):
-            if isinstance(v, lazy_value):
+            if is_lazy_value(v):
                 # a lazy value is used for several parameters at the same time, and is NOT between pytest.param()
-                argvalues[i] = LazyTuple(v, nb_params)
+                argvalues[i] = v.as_lazy_tuple(nb_params)
 
                 # TUPLE usage: we HAVE to set an id to prevent too early access to the value by _idmaker
                 # note that on pytest 2 we cannot set an id here, the lazy value wont be too lazy
@@ -1145,10 +901,10 @@ def _process_argvalues(argnames, marked_argvalues, nb_params):
                 # a fixture ref is used for several parameters at the same time
                 fixture_indices.append((i, None))
 
-            elif len(v) == 1 and isinstance(v[0], lazy_value):
+            elif len(v) == 1 and is_lazy_value(v[0]):
                 # same than above but it was in a pytest.mark
                 # valueref_indices.append((i, None))
-                argvalues[i] = LazyTuple(v[0], nb_params)  # unpack it
+                argvalues[i] = v[0].as_lazy_tuple(nb_params)  # unpack it
                 if p_ids[i] is None:
                     # force-use the id from the lazy value (do not have pytest request for it, that would unpack it)
                     p_ids[i] = v[0].get_id()
